server/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyMySQL==1.1.1
python-dotenv==1.0.1
sqids==0.4.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
waitress==3.0.0
Werkzeug==3.0.3
